refactoring observation.rs into smaller files
