+++ src/observations/mod.rs
@@ -1,1 +1,3 @@
 pub mod observations;
+pub mod trajectory_ext;
+mod observation_wrapper;
