+++ tests/reader_80col_test.rs
@@ -1,6 +1,6 @@
 use camino::Utf8Path;
 use outfit::constants::TrajectorySet;
-use outfit::observations::observations::TrajectoryExt;
+use outfit::observations::trajectory_ext::TrajectoryExt;
 
 #[test]
 fn test_80col_reader() {
